refactor: add base_model to avoid code duplication
