+++ src/apps/backend/modules/common/base_model.py
@@ -1,0 +1,13 @@
+from dataclasses import dataclass,asdict
+from typing import Any
+
+@dataclass
+class BaseModel:
+        
+    def to_bson(self) -> dict[str, Any]:
+        data = asdict(self)
+        if data.get("id") is not None:
+            data["_id"] = data.pop("id")
+        else:
+            data.pop("id", None)
+        return data
