+++ src/apps/backend/modules/password_reset_token/internal/store/password_reset_token_model.py
@@ -1,23 +1,16 @@
 from datetime import datetime
-from typing import Any, Optional
+from typing import Optional
 from bson import ObjectId
-from dataclasses import dataclass,asdict
+from dataclasses import dataclass
+from modules.common.base_model import BaseModel
 
 @dataclass
-class PasswordResetTokenModel:
+class PasswordResetTokenModel(BaseModel):
     id: Optional[ObjectId | str]
     account: ObjectId | str
     token: str
     expires_at: datetime
     is_used: bool = False
-    
-    def to_bson(self) -> dict[str, Any]:
-        data = asdict(self)
-        if data.get("id") is not None:
-            data["_id"] = data.pop("id")
-        else:
-            data.pop("id", None)
-        return data
     
     @classmethod
     def from_bson(cls, bson_data: dict) -> "PasswordResetTokenModel":
